+++ src/Caching/EventListener.php
@@ -14,7 +14,7 @@ use SplObjectStorage;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
-use Webfactory\Bundle\WfdMetaBundle\Caching\Annotation\Send304IfNotModified;
+use Webfactory\Bundle\WfdMetaBundle\Caching\Attribute\Send304IfNotModified;
 use Webfactory\Bundle\WfdMetaBundle\MetaQueryFactory;
 
 /**
@@ -93,10 +93,8 @@ class EventListener
 
     /**
      * @param $callback array A PHP callback (array) pointing to the method to reflect on.
-     *
-     * @return Send304IfNotModified|null The attribute, if found. Null otherwise.
      */
-    protected function findAttribute($callback): ?Attribute\Send304IfNotModified
+    protected function findAttribute($callback): ?Send304IfNotModified
     {
         if (!\is_array($callback)) {
             return null;
@@ -105,22 +103,8 @@ class EventListener
         $object = new ReflectionObject($callback[0]);
         $method = $object->getMethod($callback[1]);
 
-        if (\PHP_MAJOR_VERSION >= 8) {
-            $attributes = $method->getAttributes(\Webfactory\Bundle\WfdMetaBundle\Caching\Attribute\Send304IfNotModified::class);
+        $attributes = $method->getAttributes(Send304IfNotModified::class);
 
-            if ($attributes) {
-                return $attributes[0];
-            }
-        }
-
-        foreach ($this->reader->getMethodAnnotations($method) as $configuration) {
-            if ($configuration instanceof Send304IfNotModified) {
-                @trigger_error(sprintf('Using annotations to configure wfd_meta based caching on %s::%s is deprecated, use attribute-based configuration instead.', $method->class, $method->name), \E_USER_DEPRECATED);
-
-                return $configuration;
-            }
-        }
-
-        return null;
+        return $attributes ? $attributes[0] : null;
     }
 }